+++ apps/gtfs-query/query_active_trips.php
@@ -2,6 +2,12 @@
 define('APP_PATH', dirname(__FILE__));
 include(APP_PATH . '/inc/common.php');
 
+$gtfs_day = @$_GET['gtfs_day'] ?: null;
+if ($gtfs_day === null) {
+    print('error - gtfs_day is null');
+    exit(1);
+}
+
 $day = @$_GET['day'] ?: null;
 $hhmm =  @$_GET['hhmm'] ?: null;
 $from_hhmm = @$_GET['from_hhmm'] ?: null;
@@ -15,8 +21,8 @@ $parse_db_row_type = @$_GET['parse_type'] ?: 'FULL';
 
 ini_set('memory_limit', '1024M');
 
-$gtfs_controller = new GTFS_DB_Controller(APP_CONFIG, $day, $hhmm);
-$result_json = $gtfs_controller->query_active_trips($from_hhmm, $to_hhmm, $filter_agency_ids, $parse_db_row_type);
+$gtfs_controller = new GTFS_DB_Controller(APP_CONFIG, $gtfs_day);
+$result_json = $gtfs_controller->query_active_trips($day, $from_hhmm, $to_hhmm, $filter_agency_ids, $parse_db_row_type);
 
 JsonView::dump($result_json);