+++ apps/gtfs-rt-status/js/ts/models/gtfs/agency.ts
@@ -1,9 +1,0 @@
-namespace GTFS {
-    export interface Agency {
-        agency_id: string
-        agency_name: string
-        agency_url: string
-        agency_timezone: string
-        agency_lang: string
-    }
-}